verifier: combined duplicated dcap implementation
Separated duplicated ecdsa quote verification code used for dcap quote data.
