+++ deps/verifier/src/sgx/mod.rs
@@ -3,34 +3,22 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-use std::{
-    mem,
-    time::{Duration, SystemTime},
-};
-
 use anyhow::*;
 use async_trait::async_trait;
 use base64::Engine;
-use intel_tee_quote_verification_rs::{
-    quote3_error_t, sgx_ql_qv_result_t, sgx_ql_qv_supplemental_t, sgx_ql_request_policy_t,
-    sgx_qv_set_enclave_load_policy, tee_get_supplemental_data_version_and_size,
-    tee_qv_get_collateral, tee_supp_data_descriptor_t, tee_verify_quote,
-};
-use log::{debug, warn};
+use log::debug;
 use scroll::Pread;
 use serde::{Deserialize, Serialize};
 
-use crate::{regularize_data, InitDataHash, ReportData};
-
 use self::types::sgx_quote3_t;
-
 use super::{TeeEvidenceParsedClaim, Verifier};
+use crate::intel_dcap::{ecdsa_quote_verification, extend_using_custom_claims};
+use crate::{regularize_data, InitDataHash, ReportData};
 
 #[allow(non_camel_case_types)]
 mod types;
 
 mod claims;
-
 pub const QUOTE_SIZE: usize = 436;
 
 #[derive(Debug, Serialize, Deserialize)]
@@ -79,7 +67,7 @@ async fn verify_evidence(
 
     let quote_bin = base64::engine::general_purpose::STANDARD.decode(evidence.quote)?;
 
-    ecdsa_quote_verification(&quote_bin)
+    let custom_claims = ecdsa_quote_verification(&quote_bin)
         .await
         .context("Evidence's identity verification error.")?;
 
@@ -101,121 +89,10 @@ async fn verify_evidence(
         }
     }
 
-    claims::generate_parsed_claims(quote)
-}
+    let mut claim = claims::generate_parsed_claims(quote)?;
+    extend_using_custom_claims(&mut claim, custom_claims)?;
 
-async fn ecdsa_quote_verification(quote: &[u8]) -> Result<()> {
-    let mut supp_data: sgx_ql_qv_supplemental_t = Default::default();
-    let mut supp_data_desc = tee_supp_data_descriptor_t {
-        major_version: 0,
-        data_size: 0,
-        p_data: &mut supp_data as *mut sgx_ql_qv_supplemental_t as *mut u8,
-    };
-
-    // Call DCAP quote verify library to set QvE loading policy to multi-thread
-    // We only need to set the policy once; otherwise, it will return the error code 0xe00c (SGX_QL_UNSUPPORTED_LOADING_POLICY)
-    static INIT: std::sync::Once = std::sync::Once::new();
-    INIT.call_once(|| {
-        match sgx_qv_set_enclave_load_policy(
-            sgx_ql_request_policy_t::SGX_QL_PERSISTENT_QVE_MULTI_THREAD,
-        ) {
-            quote3_error_t::SGX_QL_SUCCESS => {
-                debug!("Info: sgx_qv_set_enclave_load_policy successfully returned.")
-            }
-            err => warn!(
-                "Error: sgx_qv_set_enclave_load_policy failed: {:#04x}",
-                err as u32
-            ),
-        }
-    });
-
-    match tee_get_supplemental_data_version_and_size(quote) {
-        std::result::Result::Ok((supp_ver, supp_size)) => {
-            if supp_size == mem::size_of::<sgx_ql_qv_supplemental_t>() as u32 {
-                debug!("tee_get_quote_supplemental_data_version_and_size successfully returned.");
-                debug!(
-                    "Info: latest supplemental data major version: {}, minor version: {}, size: {}",
-                    u16::from_be_bytes(supp_ver.to_be_bytes()[..2].try_into()?),
-                    u16::from_be_bytes(supp_ver.to_be_bytes()[2..].try_into()?),
-                    supp_size,
-                );
-                supp_data_desc.data_size = supp_size;
-            } else {
-                warn!("Quote supplemental data size is different between DCAP QVL and QvE, please make sure you installed DCAP QVL and QvE from same release.")
-            }
-        }
-        Err(e) => bail!(
-            "tee_get_quote_supplemental_data_size failed: {:#04x}",
-            e as u32
-        ),
-    }
-
-    // get collateral
-    let collateral = match tee_qv_get_collateral(quote) {
-        std::result::Result::Ok(c) => {
-            debug!("tee_qv_get_collateral successfully returned.");
-            Some(c)
-        }
-        Err(e) => {
-            warn!("tee_qv_get_collateral failed: {:#04x}", e as u32);
-            None
-        }
-    };
-
-    // set current time. This is only for sample purposes, in production mode a trusted time should be used.
-    //
-    let current_time = SystemTime::now()
-        .duration_since(SystemTime::UNIX_EPOCH)
-        .unwrap_or(Duration::ZERO)
-        .as_secs() as i64;
-
-    let p_supplemental_data = match supp_data_desc.data_size {
-        0 => None,
-        _ => Some(&mut supp_data_desc),
-    };
-
-    // call DCAP quote verify library for quote verification
-    let (collateral_expiration_status, quote_verification_result) = tee_verify_quote(
-        quote,
-        collateral.as_ref(),
-        current_time,
-        None,
-        p_supplemental_data,
-    )
-    .map_err(|e| anyhow!("tee_verify_quote failed: {:#04x}", e as u32))?;
-
-    debug!("tee_verify_quote successfully returned.");
-
-    // check verification result
-    match quote_verification_result {
-        sgx_ql_qv_result_t::SGX_QL_QV_RESULT_OK => {
-            // check verification collateral expiration status
-            // this value should be considered in your own attestation/verification policy
-            if collateral_expiration_status == 0 {
-                debug!("Verification completed successfully.");
-            } else {
-                warn!("Verification completed, but collateral is out of date based on 'expiration_check_date' you provided.");
-            }
-        }
-        sgx_ql_qv_result_t::SGX_QL_QV_RESULT_CONFIG_NEEDED
-        | sgx_ql_qv_result_t::SGX_QL_QV_RESULT_OUT_OF_DATE
-        | sgx_ql_qv_result_t::SGX_QL_QV_RESULT_OUT_OF_DATE_CONFIG_NEEDED
-        | sgx_ql_qv_result_t::SGX_QL_QV_RESULT_SW_HARDENING_NEEDED
-        | sgx_ql_qv_result_t::SGX_QL_QV_RESULT_CONFIG_AND_SW_HARDENING_NEEDED => {
-            warn!(
-                "Verification completed with Non-terminal result: {:x}",
-                quote_verification_result as u32
-            );
-        }
-        _ => {
-            bail!(
-                "Verification completed with Terminal result: {:x}",
-                quote_verification_result as u32
-            );
-        }
-    }
-
-    Ok(())
+    Ok(claim)
 }
 
 #[cfg(test)]
