+++ deps/verifier/src/tdx/mod.rs
@@ -6,12 +6,12 @@ use log::{debug, error, info, warn};
 use crate::{eventlog::AAEventlog, tdx::claims::generate_parsed_claim};
 
 use super::*;
+use crate::intel_dcap::{ecdsa_quote_verification, extend_using_custom_claims};
 use async_trait::async_trait;
 use base64::Engine;
 use eventlog::{CcEventLog, Rtmr};
-use quote::{ecdsa_quote_verification, parse_tdx_quote};
+use quote::parse_tdx_quote;
 use serde::{Deserialize, Serialize};
-use serde_json::{Map, Value};
 
 pub(crate) mod claims;
 pub mod eventlog;
@@ -138,20 +138,8 @@ async fn verify_evidence(
     Ok(claim)
 }
 
-pub(crate) fn extend_using_custom_claims(
-    claim: &mut TeeEvidenceParsedClaim,
-    custom: Map<String, Value>,
-) -> Result<()> {
-    let Value::Object(ref mut map) = claim else {
-        bail!("failed to extend the claim, not an object");
-    };
-    map.extend(custom);
-    Ok(())
-}
-
 #[cfg(test)]
 mod tests {
-
     use super::*;
     use std::{fs, str::FromStr};
